Feature: Reorganized advanced options for ease of use
