+++ src/Files.App/Constants.cs
@@ -198,11 +198,11 @@ namespace Files.App
 
 		public static class GitHub
 		{
-			public const string ContributorsUrl = @"https://github.com/files-community/Files/graphs/contributors";
+			public const string GitHubRepoUrl = @"https://github.com/files-community/Files";
 			public const string DocumentationUrl = @"https://files.community/docs";
-			public const string FeedbackUrl = @"https://github.com/files-community/Files/issues/new/choose";
+			public const string FeatureRequestUrl = @"https://github.com/files-community/Files/issues/new?assignees=&labels=feature+request&template=feature_request.yml";
+			public const string BugReportUrl = @"https://github.com/files-community/Files/issues/new?assignees=&labels=bug&template=bug_report.yml";
 			public const string PrivacyPolicyUrl = @"https://github.com/files-community/Files/blob/main/Privacy.md";
-			public const string ReleaseNotesUrl = @"https://github.com/files-community/Files/releases";
 			public const string SupportUsUrl = @"https://github.com/sponsors/yaira2";
 		}
 	}
