+++ src/Files.App/Views/SettingsPages/Advanced.xaml.cs
@@ -2,9 +2,9 @@ using Microsoft.UI.Xaml.Controls;
 
 namespace Files.App.SettingsPages
 {
-	public sealed partial class Experimental : Page
+	public sealed partial class Advanced : Page
 	{
-		public Experimental()
+		public Advanced()
 		{
 			InitializeComponent();
 		}
